+++ app/components/preprint-form-project-select.js
@@ -30,7 +30,7 @@ import {stripDiacritics} from 'ember-power-select/utils/group-utils';
  *     startState=_State.START
  *     existingState=existingState
  *     _existingState=_existingState
- *     nodeTitle=nodeTitle
+ *     title=title
  *     currentUser=user
  *     selectedFile=selectedFile
  *     hasFile=hasFile