+++ tests/unit/controllers/content/index-test.js
@@ -69,11 +69,11 @@ test('twitterHref computed property', function (assert) {
     const ctrl = this.subject();
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             title: 'test title'
         });
 
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         const location = encodeURIComponent(window.location.href);
 
@@ -145,12 +145,12 @@ test('linkedinHref computed property', function (assert) {
     const ctrl = this.subject();
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             title: 'test title',
             description: 'test description'
         });
 
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         const location = encodeURIComponent(window.location.href);
 
@@ -168,11 +168,11 @@ test('emailHref computed property', function (assert) {
     const ctrl = this.subject();
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             title: 'test title'
         });
 
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         const location = encodeURIComponent(window.location.href);
 
@@ -190,11 +190,11 @@ test('hasTag computed property', function (assert) {
     const ctrl = this.subject();
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             tags: []
         });
 
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         assert.strictEqual(
             ctrl.get('hasTag'),
@@ -203,11 +203,11 @@ test('hasTag computed property', function (assert) {
     });
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             tags: ['a', 'b', 'c']
         });
 
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         assert.strictEqual(
             ctrl.get('hasTag'),
@@ -225,13 +225,11 @@ skip('authors computed property', function (assert) {
     const ctrl = this.subject();
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             id: 'abc12'
         });
 
-        ctrl.setProperties({
-            node
-        });
+        ctrl.set('model', preprint);
 
         // TODO figure out how to test with at least one contributor
         ctrl.get('authors')
@@ -335,11 +333,10 @@ test('hasShortenedDescription computed property', function (assert) {
     const ctrl = this.subject();
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             description: 'Lorem ipsum'
         });
-
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         assert.strictEqual(
             ctrl.get('hasShortenedDescription'),
@@ -348,11 +345,11 @@ test('hasShortenedDescription computed property', function (assert) {
     });
 
     Ember.run(() => {
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             description: 'Lorem ipsum'.repeat(35)
         });
 
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         assert.strictEqual(
             ctrl.get('hasShortenedDescription'),
@@ -391,11 +388,11 @@ test('description computed property', function (assert) {
         const input = 'Lorem ipsum dolor sit amet, atqui elitr id vim, at clita facilis tibique ius, ad pro stet accusam. Laudem essent commune ea vix. Duis hendrerit complectitur usu eu, ei nam ullum accusamus inciderint, has appetere assueverit te. An pro maiorum alienum voluptatibus, mei adhuc docendi prodesset in. Ut vel mundi atomorum quaerendum, cu per autem menandri consequat, tantas dictas quodsi nec eu. Ornatus forensibus vituperatoribus id vix.';
         const expected ='Lorem ipsum dolor sit amet, atqui elitr id vim, at clita facilis tibique ius, ad pro stet accusam. Laudem essent commune ea vix. Duis hendrerit complectitur usu eu, ei nam ullum accusamus inciderint, has appetere assueverit te. An pro maiorum alienum voluptatibus, mei adhuc docendi prodesset in. Ut vel mundi atomorum quaerendum, cu per autem';
 
-        const node = store.createRecord('node', {
+        const preprint = store.createRecord('preprint', {
             description: input
         });
 
-        ctrl.setProperties({node});
+        ctrl.set('model', preprint);
 
         assert.strictEqual(
             ctrl.get('description'),