+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - Default auto-expansion on top level subjects when there are no more than 3.
 - Check to use `facebookAppId` if branded providers have an app id
@@ -12,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - 'Powered by Preprints' link stay on current server
 - Searching for contributors also searches their social links (such as their twitter handle)
+- Pull more preprint data (`title`, `description`, `tags`, etc.) from the preprint model instead of the node model.
 
 ### Fixed
 - Skipped tests in `convert-or-copy-project-test` and `supplementary-file-browser` to run properly
@@ -23,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.117.4] - 2018-02-21
 ### Fixed
 - Increase taxonomies page size from 100 to 150
+
 
 ## [0.117.3] - 2018-02-20
 ### Changed